Fixing Database Migration
