+++ backend/framework/alembic/versions/7c9ef8e9c5b9_damurill_generate_monsters.py
@@ -1,0 +1,42 @@
+"""damurill_generate_monsters
+
+Revision ID: 7c9ef8e9c5b9
+Revises: aa75144a150e
+Create Date: 2025-08-27 10:24:42.407891
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '7c9ef8e9c5b9'
+down_revision: Union[str, Sequence[str], None] = 'aa75144a150e'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('dmuri_monsters',
+    sa.Column('index', sa.String(length=100), nullable=False),
+    sa.Column('data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('index')
+    )
+    op.create_table('dmuri_monsters_list',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('data', sa.JSON(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('dmuri_monsters_list')
+    op.drop_table('dmuri_monsters')
+    # ### end Alembic commands ###
