--- backend/framework/alembic/versions/d38138823f33_generate_monster_modelcls.py
@@ -1,54 +1,0 @@
-"""Generate Monster Modelcls
-
-Revision ID: d38138823f33
-Revises: 1d0f25916ca5
-Create Date: 2025-08-27 02:31:15.881883
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'd38138823f33'
-down_revision: Union[str, Sequence[str], None] = '1d0f25916ca5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('dmrui_monsters_list',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('data', sa.JSON(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('dmuri_motivational_phrases',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('phrase', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.drop_table('motivational_phrases')
-    op.drop_table('monsters_list')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('monsters_list',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('data', postgresql.JSON(astext_type=sa.Text()), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name=op.f('monsters_list_pkey'))
-    )
-    op.create_table('motivational_phrases',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('phrase', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name=op.f('motivational_phrases_pkey'))
-    )
-    op.drop_table('dmuri_motivational_phrases')
-    op.drop_table('dmrui_monsters_list')
-    # ### end Alembic commands ###
